misrepo31 added

+++ src/main/java/com/aristowebapi/dto/BranchMisRepo8.java
@@ -7,7 +7,7 @@ public interface BranchMisRepo8 {
 	public int  getFs();
 	public int getTerr_cd();
 	public String getTer_name();
-	public int getSales();
+	public long getSales();
 	public long getSales_val();
 	public String getPname();
     public String getSmname();
